+++ 02-electrondensity/src/Timer.cpp
@@ -10,7 +10,7 @@ void Timer::stop()
     t_end = std::chrono::high_resolution_clock::now();
 }
 
-double Timer::getDuration()
+float Timer::getDuration()
 {
     std::chrono::microseconds time_elapsed = std::chrono::duration_cast<std::chrono::microseconds>(t_end - t_start);
     return time_elapsed.count();
